Update README and index pages
Updated the project's README and index pages to help give readers
information on what the project is about and where to get started.

The README now includes a link to the latest released version and
build instructions.

The index page is updated to include sections for the overview and
design documents along with some getting started information.