Add changelog / clean up HTML build directory
A changelog is added so others know what changes have occurred between
revisions.

The zip file that gets generated from the HTML build directory is then
moved up a level. This ensures one can copy and paste the entire HTML
directory to a web server without first having to delete or move the
zip file. Additionally, the zip is not easier to find.